Update appengine API.
The backend is using an outdated version of the appengine API. The new
version is hosted at "google.golang.org/appengine". Unlike the current
"appengine" package, the new version is supported by third party tools.
It has also been cleaned up a bit and is the version referred to by the
current App Engine documentation.

Unfortunately, not a lot seems to have been written about the
transition. Some information is here:

https://code.google.com/p/googleappengine/issues/detail?id=11670

and here:

https://github.com/golang/appengine/blob/0f8c70efe6e140f4028f4182b6ad4f4822d63330/README.md

Most of this commit was written automatically by the "aefix" tool that
the Go team created to help people like us upgrade to the new API. On
top of that, I also had to add logging functions in
src/sessions/context.go. The new API uses a dedicated log package rather
than having logging functions on the context variable. But since we wrap
the context in our own struct, aefix didn't know how to update it.
Fortunately, writing a custom wrapper is straightforward.