Vendor Go dependencies.
This is a very common approach to managing dependencies in Go. A lot has
been written about it online. See e.g.

https://blog.gopheracademy.com/advent-2015/vendor-folder/
https://docs.google.com/document/d/1Bz5-UB7g2uPBdOx-rw5t9MxJwkfpx90cqG9AFL0JAYo/edit

The short justification is that it pins all of our dependencies to
specific versions, which means the build is the same for everyone.
Without vendoring dependencies, everyone who builds the project gets
something slightly different depending on what the version of every
library was at the time they first built the project.

This also makes installation much easier, since we don't have to
separately download dependencies.

There are multiple tools for managing Go dependencies. One of the most
popular ones is github.com/tools/godep. Unfortunately, godep had some
problems understanding our appengine dependencies, so this commit uses
github.com/kardianos/govendor instead.

This commit also deletes src/googleauth, which is both unused and
depends on a library that no longer exists on the internet.

NOTE: Building the project with these dependencies will require a change
in the standard way we set up our dev environments. Our current approach
deviates from the standard way of using GOPATH, and no longer works with
vendored dependencies. A fix for this is coming in the next commit.