Make scripts compatible with living in $GOPATH/src/zanaduu3
compile_css.sh and style_js.sh expect to live under a directory called
"default". But the Go packages can't build if they are under a directory
called "default", which we currently solve with a symlink.

Some Go tools break with the symlink solution, though. A better solution
is to have the project live in $GOPATH/src/zanaduu3 and update these
scripts.

Also delete gcloud_bootstrap.sh, which has beliefs about the default
directory, but looks unused and not worth updating.