Use String.prototype.localeCompare instead of '<'
The compare function for Array.prototype.sort should return -1, 0, or 1
for a < b, a == b, and a > b, respectively. This comparison currently
returns true or false, which is presumably type cast to 0 or 1. This
means that it can't ever say that a < b, which means it is not sorting
correctly.