Switch to webpack and TypeScript.
Webpack interprets our javascript files as distinct modules that run in
their own namespace, and which must explicitly import their dependencies
and export the identifiers other files need.

Arbital's javascript currently all lives in a global namespace, and
doesn't explicitly declare its dependencies. This means that if we start
using them as modules, they will fail at runtime when the global
namespace identifiers they are expecting are no longer defined.

To avoid this failure, we need a way to statically verify that all our
dependencies are declared, and that those declarations actually exist.
Unfortunately, modern static analysis in javascript leaves something to
be desired. I am not aware of any tool that can do this verification on
plain javascript.

Enter TypeScript. TypeScript is a superset of javascript that adds
gradual typing. And it has a compiler that does type checking and
understands modules. The TypeScript compiler will do the dependency
verification we want, plus add additional type safety.

A downside of doing things this way is that this change must be very
large. We are adding webpack and causing our entire javascript codebase
to type check and declare its dependencies all at once.

Note that this commit requires installing `typings` and `webpack`,
running `typings install` in src/site/static/js. Then every time you
change a TypeScript file or an SCSS file, running `webpack` in
src/site/webpack. This dev flow is not ideal, and will be smoothed out
in another commit before this is merged to master.