Fix templating for dynamic pages
There are currently two places where we set data for the Go template in
tmp/dynamicPage.tmpl: pageHandlerWrapper and pageRenderer.

pageHandlerWrapper currently wraps three kinds of pages: pagePage,
dynamicPage, and verifyEmailPage. dynamicPage and verifyEmailPage do not
set any data for the Go template, so pageHandlerWrapper sets it on their
behalf. pagePage does set data, so pageHandlerWrapper was not setting
any data for those pages.

The fact that pageHandlerWrapper does not set data for pagePages  was
not obvious to me when I integrated webpack, and it wasn't obvious to
whomever added VersionId for asset versioning. As a result, the data
needed for webpack and versioning was never set for pagePages. What that
means in practice is that Arbital articles that were loaded from scratch
(as opposed to navigated to from the Arbital home page), failed to have
asset versioning and failed to work correctly with webpack.

This commit fixes this problem by merging the data in pageHandlerWrapper
with the data from pagePage.