Patch AppEngine to not watch node_modules/** for changes
node_modules/ contains a lot of files, and AppEngine doesn't need to
care about any of them. But by default AppEngine watches them all for
changes. AppEngine also sets a limit on the number of directories it
watches, and watching node_modules causes it to exceed this limit. This
means it can fail to watch files we actually care about because it is
instead watching files we don't care about.

There is currently no supported way to configure AppEngine to not watch
some files. There is discussion about using the `skip_files` app.yaml
configuration for this:

    https://code.google.com/p/googleappengine/issues/detail?id=9300

...but it has been outstanding for over three years and seems unlikely to
be merged any time soon.

This change takes the approach of patching AppEngine. Other people have
done the same thing:

    http://johanndutoit.net/app-engine-too-many-files/
    https://github.com/Khan/frankenserver

The patch in this commit is small, and I have included a script,
`patch_appengine.sh`, that will apply it automatically.

This commit also adds node_modules to the skip_files configuration in
app.yaml to prevent AppEngine from uploading all of those files when we
deploy.