Auto-run `npm install` when packages or typings are updated.
Now when npm packages or typings files are changed, the upgrade process
will be a little easier. Developers will still need to restart stage.sh,
though. It might be nice to automatically do that in the future, as
well.

This commit also adds a dependency on
[gorun](https://github.com/jeremyschlatter/gorun), which allows us to
write scripts in Go rather than bash.